+++ lib/store/my_user.dart
@@ -46,7 +46,6 @@ import '/provider/gql/graphql.dart';
 import '/util/backoff.dart';
 import '/util/event_pool.dart';
 import '/util/log.dart';
-import '/util/multipart_file.dart';
 import '/util/new_type.dart';
 import '/util/obs/obs.dart';
 import '/util/platform_utils.dart';
@@ -681,7 +680,7 @@ class MyUserRepository extends DisposableInterface
     if (file != null) {
       await file.ensureCorrectMediaType();
 
-      upload = await multipartFromNativeFile(file);
+      upload = await file.toMultipartFile();
     }
 
     final UserAvatar? avatar = myUser.value?.avatar;
@@ -751,7 +750,7 @@ class MyUserRepository extends DisposableInterface
     if (file != null) {
       await file.ensureCorrectMediaType();
 
-      upload = await multipartFromNativeFile(file);
+      upload = await file.toMultipartFile();
     }
 
     final UserCallCover? callCover = myUser.value?.callCover;
